Introduced module implementing all test helper methods
this module is included by all tests within the test_helper
reduced code duplications
