+++ test/helper_methods.rb
@@ -1,0 +1,13 @@
+module HelperMethods
+  # allow to fetch puts outputs
+  def fetch_stdout(&block)
+    begin
+      old_stdout = $stdout
+      $stdout = StringIO.new('','w')
+      yield block
+      $stdout.string
+    ensure
+      $stdout = old_stdout
+    end
+  end
+end
