+++ test/test_layer.rb
@@ -2,18 +2,6 @@ require "test_helper"
 
 class TestLayer < Minitest::Test
 
-  # allow to fetch puts outputs
-  def fetch_stdout(&block)
-    begin
-      old_stdout = $stdout
-      $stdout = StringIO.new('','w')
-      yield block
-      $stdout.string
-    ensure
-      $stdout = old_stdout
-    end
-  end
-
   class ANewDrawableL
     def initialize(id="")
       @id = id
